Create Lab5.v
The following Verilog file has multiple responsibilities. Firstly, the code is responsible for converting the 4 bit input, which is activated from the pushbuttons, to a 7 bit value, which represents the corresponding BCD value. Secondly, the file consists of a case statement that reflects the behavior of incrementing by 1, incrementing by 10, and resetting the score.